Now you can specify the folder to store the documents

+++ BVMFCrawler.py
@@ -1,5 +1,6 @@
 import sys
 import requests
+import os
 from bs4 import BeautifulSoup
 
 DOCUMENTS_URL = 'http://bvmf.bmfbovespa.com.br/Fundos-Listados/FundosListadosDetalhe.aspx?Sigla={}' \
@@ -12,9 +13,12 @@ OUTROSDOCS_URL = 'http://bvmf.bmfbovespa.com.br/Fundos-Listados/FundosListadosDe
                  '&tipoFundo=Imobiliario&aba=subAbaOutrosDocumentos&idioma=pt-br'
 
 
-def download_file(url, local_filename):
+# given a PDF url, download the document to the folder specifiec
+def download_file(url, local_filename, folder):
+    path = os.path.join(folder, local_filename.replace('/', ' ') + '.pdf')
+    if os.path.isfile(path):
+        return
     print("Fazendo download de {}".format(local_filename))
-    path = local_filename.replace('/', ' ') + '.pdf'
     r = requests.get(url, stream=True)
     with open(path, 'wb') as f:
         for chunk in r.iter_content(chunk_size=1024):
@@ -23,23 +27,39 @@ def download_file(url, local_filename):
     return local_filename
 
 
-def get_documents(url):
+# connect to BVMF pages and get all downloads links
+def get_documents(url, folder):
     html_page = requests.get(url)
     soup = BeautifulSoup(html_page.content, "html.parser")
     for link in soup.findAll('a'):
         doc = link.get('href')
         if 'FormConsultaPdfDocumentoFundos' in doc:
-            download_file(doc, link.string)
+            download_file(doc, link.string, folder)
+
+
+# creates all folders that will store the documents
+def create_folders(rootfolder):
+    fullpath = os.path.join(rootfolder, "DOCUMENTS")
+    if not os.path.exists(fullpath):
+        os.makedirs(fullpath)
+    fullpath = os.path.join(rootfolder, "DEMOFINANCEIRAS")
+    if not os.path.exists(fullpath):
+        os.makedirs(fullpath)
+    fullpath = os.path.join(rootfolder, "OUTROSDOCS")
+    if not os.path.exists(fullpath):
+        os.makedirs(fullpath)
 
 
 def main():
     fundo = sys.argv[1]
+    rootfolder = sys.argv[2]
+    create_folders(rootfolder)
     url = DOCUMENTS_URL.format(fundo)
-    get_documents(url)
+    get_documents(url, os.path.join(rootfolder, "DOCUMENTS"))
     url = DEMOFINANCEIRAS_URL.format(fundo)
-    get_documents(url)
+    get_documents(url, os.path.join(rootfolder, "DEMOFINANCEIRAS"))
     url = OUTROSDOCS_URL.format(fundo)
-    get_documents(url)
+    get_documents(url, os.path.join(rootfolder, "OUTROSDOCS"))
 
 
 if __name__ == '__main__':
